auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0
supabase==2.5.1
python-dotenv==1.0.1
email-validator==2.2.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
structlog>=23.2.0
httpx>=0.25.0
gunicorn>=21.2.0
# 监控相关依赖
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0
opentelemetry-exporter-jaeger-thrift>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-propagator-b3>=1.21.0
opentelemetry-propagator-jaeger>=1.21.0
sentry-sdk[fastapi]>=1.38.0
psutil>=5.9.0